split ChunkResolver/ChunkedArrayResolver files
