+++ cpp/src/arrow/compute/kernels/vector_sort_internal.h
@@ -24,6 +24,7 @@
 
 #include "arrow/array.h"
 #include "arrow/compute/api_vector.h"
+#include "arrow/chunked_array_internal.h"
 #include "arrow/compute/kernels/chunked_internal.h"
 #include "arrow/type.h"
 #include "arrow/type_traits.h"
